Code/requirements.txt
streamlit
nltk
gensim
pyvi
networkx
textract
docx
requests
bs4
python-docx
pdfminer.six